feat(typechecking): Added primitive type checking framework
This change utilizes the applyWithProgram function added to AbstractRule
in the cli. Currently files in the same "project" must all be specified
on the command line. This should be changed to read the tsconfig or
included in the codelyzer config in the future. The reporter / generator
pattern was changed to a `.fix.ts` file for the moment.